Move RPC change classification out of wire codecs
