+++ app/ts/background/activeSettings.ts
@@ -3,7 +3,8 @@ import type { EthereumClientService } from '../simulation/services/EthereumClien
 import type { ResetSimulationServices, SimulationServices } from '../simulation/serviceLifecycle.js'
 import type { TokenPriceService } from '../simulation/services/priceEstimator.js'
 import type { SigningAddressPreference } from '../types/signerTypes.js'
-import { getRpcNetworkChange, type RpcNetwork } from '../types/rpc.js'
+import { getRpcNetworkChange } from '../utils/rpcNetworkChange.js'
+import type { RpcNetwork } from '../types/rpc.js'
 import type { WebsiteTabConnections } from '../types/user-interface-types.js'
 import { Semaphore } from '../utils/semaphore.js'
 import type { WebsiteAccessUpdate } from './accessManagement.js'
