+++ app/ts/background/windows/changeChain.ts
@@ -6,7 +6,8 @@ import type { WebsiteTabConnections } from '../../types/user-interface-types.js'
 import { changeActiveRpc } from '../activeSettings.js'
 import { getHtmlFile, getSocketFromPort, sendPopupMessageToOpenWindows } from '../backgroundUtils.js'
 import { getChainChangeConfirmationPromise, getRpcNetworkForChain, promoteRpcAsPrimary, setChainChangeConfirmationPromise } from '../storageVariables.js'
-import { getRpcNetworkChange, type RpcNetwork } from '../../types/rpc.js'
+import { getRpcNetworkChange } from '../../utils/rpcNetworkChange.js'
+import type { RpcNetwork } from '../../types/rpc.js'
 import { type InterceptedRequest, type UniqueRequestIdentifier, doesUniqueRequestIdentifiersMatch } from '../../utils/requests.js'
 import { replyToInterceptedRequest } from '../messageSending.js'
 import type { SwitchEthereumChainParams } from '../../types/JsonRpc-types.js'
