+++ app/ts/components/hooks/useLiveSimulationHomeData.ts
@@ -3,7 +3,8 @@ import { MessageToPopup, type HomePageBootstrap, type UpdateHomePage, type Setti
 import type { RpcConnectionStatus, TabIconDetails, TabState } from '../../types/user-interface-types.js'
 import { PASSTHROUGH_STATE, type BlockTimeManipulation, type CompleteVisualizedSimulation, type NamedTokenId, type ResolvedSimulationResults, type ResolvedSimulationState, type SimulationResultState, type SimulationUpdatingState, type TokenPriceEstimate, type VisualizedSimulationState, toResolvedSimulationResults } from '../../types/visualizer-types.js'
 import type { AddressBookEntries, AddressBookEntry } from '../../types/addressBookTypes.js'
-import { getRpcNetworkChange, type RpcEntries, type RpcNetwork } from '../../types/rpc.js'
+import { getRpcNetworkChange } from '../../utils/rpcNetworkChange.js'
+import type { RpcEntries, RpcNetwork } from '../../types/rpc.js'
 import type { WebsiteAccessArray } from '../../types/websiteAccessTypes.js'
 import type { EnrichedRichListElement, UnexpectedErrorOccured } from '../../types/interceptor-reply-messages.js'
 import { PopupMessageReplyRequests } from '../../types/interceptor-reply-messages.js'
