+++ app/ts/types/rpc.ts
@@ -49,13 +49,3 @@ export const RpcNetwork = funtypes.Union(
 		minimized: funtypes.Literal(true),
 	})
 )
-
-// Chain changes require wallet approval; endpoint changes replace services; metadata changes only update the selection.
-export function getRpcNetworkChange(previous: RpcNetwork | undefined, next: RpcNetwork) {
-	const chainChanged = previous?.chainId !== next.chainId
-	return {
-		chainChanged,
-		endpointChanged: chainChanged || previous?.httpsRpc !== next.httpsRpc,
-		selectionChanged: previous === undefined || JSON.stringify(RpcNetwork.serialize(previous)) !== JSON.stringify(RpcNetwork.serialize(next)),
-	}
-}
